requirements.txt
requests
beautifulsoup4
pandas
google-cloud-bigquery
openpyxl
